requirements.txt
streamlit
streamlit-authenticator
openai
pandas
requests
beautifulsoup4
pdfplumber
fpdf
python-docx
PyMuPDF
